pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.49.1",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "sentencepiece>=0.2.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
